# docs/openapi.yaml
openapi: 3.0.3
info:
  title: KCT Knowledge API
  description: |
    Comprehensive menswear styling API providing intelligent outfit validation, color recommendations, 
    and fashion intelligence powered by extensive knowledge bank data.
    
    ## Features
    - **Advanced Validation Engine**: Multi-category rule validation with confidence scoring
    - **Color Harmony Analysis**: Sophisticated color relationship analysis
    - **Formality Matching**: Occasion and dress code compliance validation  
    - **Seasonal Appropriateness**: Fabric and color seasonality validation
    - **Style Profile Matching**: Personalized recommendations based on customer profiles
    - **Venue Requirements**: Location-specific styling requirements
    - **Pattern Mixing Rules**: Advanced pattern compatibility analysis
    - **Trending Intelligence**: Real-time fashion trend analysis
    
    ## Authentication
    All endpoints require API key authentication via the `X-API-Key` header.
    
    ## Rate Limits
    - Standard plan: 1000 requests/hour
    - Premium plan: 10000 requests/hour
    - Enterprise plan: Unlimited
    
    ## Response Format
    All responses follow a consistent format with success/error status, data payload, and metadata.
    
  version: 2.0.0
  contact:
    name: KCT API Support
    email: api-support@kctmenswear.com
    url: https://api.kctmenswear.com/support
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
...
servers:
...
security:
  - ApiKeyAuth: []
paths:
  /colors:
    get:
      summary: List all colors with metadata and relationships
      description: |
        Retrieves comprehensive color data including color families, universal rules, 
        and trending color information from the knowledge bank.
      tags:
        - Colors
      responses:
        '200':
          description: Successful response with color data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              example:
                success: true
                data:
                  color_families:
                    neutrals:
                      colors: ["white", "black", "grey", "charcoal"]
                      complement_with: "any color family"
                      tie_suggestions: ["burgundy", "navy", "silver"]
                      trending: "stable"
                  universal_rules:
                    white_shirt_universal:
                      works_with: "all suit colors"
                      confidence: 0.95
                      note: "Always appropriate and versatile"
                  trending:
                    rising: ["sage_green", "burgundy", "midnight_blue"]
                    stable: ["navy", "charcoal", "white"]
                    declining: ["tan", "light_blue"]
                  total_colors: 47
                  metadata:
                    version: "2.0.0"
                    last_updated: "2024-01-15T10:30:00Z"
                    description: "Comprehensive color relationship data for menswear styling"
                error: null
                timestamp: "2024-01-15T10:30:00Z"
        '500':
          $ref: '#/components/responses/InternalServerError'
  /colors/{color}/relationships:
    get:
      summary: Get matching colors for suits/shirts/ties
      description: |
        Retrieves complementary color relationships for a specific color,
        including confidence scores and seasonal recommendations.
      tags:
        - Colors
      parameters:
        - name: color
          in: path
          required: true
          description: Color name (e.g., 'navy', 'charcoal', 'burgundy')
          schema:
            type: string
            example: navy
      responses:
        '200':
          description: Color relationships found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              example:
                success: true
                data:
                  color: "navy"
                  normalized_color: "navy"
                  relationships:
                    perfect_matches:
                      shirts: ["white", "light_blue"]
                      ties: ["burgundy", "silver", "gold"]
                      confidence: 0.95
                    good_matches:
                      shirts: ["pink", "lavender"]
                      ties: ["navy", "forest_green"]
                      confidence: 0.85
                  confidence_scores:
                    shirt_matches: 0.95
                    tie_matches: 0.90
                    overall_confidence: 0.92
                  seasonal_recommendations:
                    spring: ["light_blue_shirt", "coral_tie"]
                    summer: ["white_shirt", "sage_green_tie"]
                  formality_notes: "versatile for business to formal occasions"
                  metadata:
                    source: "KCT Knowledge Bank"
                    last_updated: "2024-01-15T10:30:00Z"
                error: null
                timestamp: "2024-01-15T10:30:00Z"
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /combinations/validate:
    post:
      summary: Validate outfit combinations with comprehensive analysis
      description: |
        Performs comprehensive validation of outfit combinations using advanced
        fashion rules engine with confidence scoring and detailed feedback.
      tags:
        - Validation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CombinationValidationRequest'
            example:
              suit_color: "navy"
              suit_fabric: "wool"
              shirt_color: "white"
              shirt_pattern: "solid"
              tie_color: "burgundy"
              tie_pattern: "solid"
              occasion: "business_formal"
              season: "fall"
              venue_type: "office"
              customer_profile: "classic_conservative"
              formality_required: 7
      responses:
        '200':
          description: Validation completed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              example:
                success: true
                data:
                  overall_score: 92
                  confidence: 0.95
                  validation_passed: true
                  category_scores:
                    color_harmony: 0.95
                    formality_matching: 0.90
                    seasonal_appropriateness: 0.88
                    style_consistency: 0.92
                  rule_results:
                    - rule_id: "CH001"
                      rule_name: "Color Harmony"
                      category: "color_harmony"
                      passed: true
                      confidence: 0.95
                      severity: "success"
                      message: "Excellent color harmony"
                      reasoning: "Navy, white, and burgundy create classic sophisticated combination"
                      recommendation: "Perfect color choice"
                      score_impact: 5
                  violations: []
                  warnings: []
                  suggestions: []
                  alternatives: []
                  improvement_suggestions: []
                  metadata:
                    rules_processed: 15
                    processing_time_ms: 245
                    knowledge_bank_version: "2.0.0"
                    validated_at: "2024-01-15T10:30:00Z"
                error: null
                timestamp: "2024-01-15T10:30:00Z"
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /recommendations:
    post:
      summary: AI-powered outfit recommendations
      description: |
        Generates personalized outfit recommendations based on customer profile,
        occasion, and preferences using advanced AI algorithms.
      tags:
        - Recommendations
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecommendationRequest'
            example:
              suit_color: "navy"
              customer_profile: "modern_adventurous"
              occasion: "cocktail"
              season: "spring"
              formality_level: 6
              age: 32
              occupation: "marketing_executive"
      responses:
        '200':
          description: Recommendations generated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              example:
                success: true
                data:
                  primary_recommendations:
                    - suit_color: "navy"
                      shirt_color: "light_blue"
                      tie_color: "coral"
                      rank: 1
                      ai_confidence: 0.95
                      personalization_score: 0.92
                      trending_factor: 0.88
                      reasoning: "Perfect balance of classic and contemporary for modern adventurous profile"
                  alternative_options:
                    - suit_color: "navy"
                      shirt_color: "white"
                      tie_color: "forest_green"
                      confidence: 0.87
                      reasoning: "Alternative with nature-inspired accent color"
                  style_insights:
                    detected_profile: "modern_adventurous"
                    confidence: 0.90
                    key_characteristics:
                      color_preferences: ["contemporary_colors", "accent_experimentation"]
                      pattern_tolerance: "moderate"
                      adventure_level: "medium"
                    personalization_factors:
                      - "Color preferences"
                      - "Occasion requirements"
                      - "Age-appropriate styling"
                  upsell_opportunities:
                    - "Premium silk tie upgrade"
                    - "Seasonal pocket square addition"
                  seasonal_highlights:
                    season: "spring"
                    trending_colors: ["sage_green", "coral", "powder_blue"]
                    seasonal_tips:
                      - "Perfect for spring events"
                      - "Seasonal color palette optimized"
                  metadata:
                    generated_at: "2024-01-15T10:30:00Z"
                    algorithm_version: "2.0.0"
                    knowledge_bank_version: "2.0.0"
                    total_combinations_analyzed: 1247
                error: null
                timestamp: "2024-01-15T10:30:00Z"
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for authentication
  schemas:
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
          description: Indicates if the request was successful
        data:
          type: object
          description: Response data payload
          nullable: true
        error:
          type: string
          description: Error message if request failed
          nullable: true
        timestamp:
          type: string
          format: date-time
          description: Response timestamp
      required:
        - success
        - timestamp
...
    RecommendationRequest:
      type: object
      required:
        - suit_color
      properties:
        suit_color:
          $ref: '#/components/schemas/SuitColor'
        customer_profile:
          $ref: '#/components/schemas/CustomerProfile'
        occasion:
          $ref: '#/components/schemas/Occasion'
        season:
          $ref: '#/components/schemas/Season'
        formality_level:
          type: integer
          minimum: 1
          maximum: 10
        age:
          type: integer
          minimum: 18
          maximum: 100
        occupation:
          type: string
          example: marketing_executive
...
    SuitColor:
      type: string
      enum:
        - navy
        - charcoal
        - light_grey
        - burgundy
        - black
        - tan
        - light_blue
        - sage_green
        - hunter_green
        - powder_blue
        - midnight_blue
        - white
        - brown
      example: navy
...
    Season:
      type: string
      enum: [spring, summer, fall, winter]
      example: fall
...
  responses:
    BadRequest:
      description: Bad request - invalid parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
          example:
            success: false
            data: null
            error: "suit_color, shirt_color, and tie_color are required"
            timestamp: "2024-01-15T10:30:00Z"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
          example:
            success: false
            data: null
            error: "Resource not found"
            timestamp: "2024-01-15T10:30:00Z"
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
          example:
            success: false
            data: null
            error: "Internal server error"
            timestamp: "2024-01-15T10:30:00Z"
...